Fix GetCurrentShopAsync and add more classes and properties.
'GetCurrentAsync' had an issue with entries not displaying, but after adding required classes, the problem is now resolved. when using (`/shop`), it will list all bundles that are available.

In FortniteModule.cs, a minor fix was changed to set the ephermal to false, so that when the command is used, it is visible to everyone.